todo/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.12.0
pytz==2019.3
sqlparse==3.0.0
djangorestframework-jwt==1.11.0